Add exists_table to catalog trait
