+++ query/src/catalogs/catalog.rs
@@ -14,6 +14,7 @@
 
 use std::sync::Arc;
 
+use common_exception::ErrorCode;
 use common_exception::Result;
 use common_meta_types::CreateDatabaseReply;
 use common_meta_types::CreateDatabaseReq;
@@ -27,6 +28,7 @@ use common_meta_types::TableIdent;
 use common_meta_types::TableInfo;
 use common_meta_types::TableMeta;
 use common_meta_types::UpsertTableOptionReply;
+use dyn_clone::DynClone;
 
 use crate::catalogs::Database;
 use crate::catalogs::Table;
@@ -39,7 +41,7 @@ use crate::datasources::table_func_engine::TableArgs;
 /// When we create a new database, we first to get the engine from the registered engines,
 /// and use the engine to create them.
 #[async_trait::async_trait]
-pub trait Catalog {
+pub trait Catalog: DynClone + Send + Sync {
     // Get all the databases.
     async fn get_databases(&self) -> Result<Vec<Arc<dyn Database>>>;
 
@@ -49,6 +51,7 @@ pub trait Catalog {
     // Get one table by db and table name.
     async fn get_table(&self, db_name: &str, table_name: &str) -> Result<Arc<dyn Table>>;
 
+
     async fn get_tables(&self, db_name: &str) -> Result<Vec<Arc<dyn Table>>>;
 
     async fn get_table_meta_by_id(&self, table_id: MetaId) -> Result<(TableIdent, Arc<TableMeta>)>;
@@ -56,6 +59,20 @@ pub trait Catalog {
     async fn create_table(&self, req: CreateTableReq) -> Result<()>;
 
     async fn drop_table(&self, req: DropTableReq) -> Result<DropTableReply>;
+
+    // Check a db.table is exists or not.
+    async fn exists_table(&self, db_name: &str, table_name: &str) -> Result<bool> {
+        match self.get_table(db_name, table_name).await {
+            Ok(_) => Ok(true),
+            Err(err) => {
+                if err.code() == ErrorCode::UnknownTableCode() {
+                    Ok(false)
+                } else {
+                    Err(err)
+                }
+            }
+        }
+    }
 
     /// Build a `Arc<dyn Table>` from `TableInfo`.
     fn build_table(&self, table_info: &TableInfo) -> Result<Arc<dyn Table>>;
@@ -82,5 +99,16 @@ pub trait Catalog {
 
     async fn drop_database(&self, req: DropDatabaseReq) -> Result<()>;
 
-    async fn exists_database(&self, db_name: &str) -> Result<bool>;
+    async fn exists_database(&self, db_name: &str) -> Result<bool> {
+        match self.get_database(db_name).await {
+            Ok(_) => Ok(true),
+            Err(err) => {
+                if err.code() == ErrorCode::UnknownDatabaseCode() {
+                    Ok(false)
+                } else {
+                    Err(err)
+                }
+            }
+        }
+    }
 }
